requirements.txt
flask
edge-tts
gunicorn
python-docx
PyPDF2